The missing value backingstore table is now *always* created. (re #86)
Previously, the missing value backing store was not created unless there was at least one missing value.  Unfortunately, the nested (SELECT) makes into the query regardless of whether there is an actual missing value or not.  As a result, the backend database complains when it can't find the backing store, breaking things.

This doesn't actually close the bug report -- there's a second part to this regression.  As already noted, SQLite silently fails when casting non-numerics to REALs.  Will, the rest of this is all you.