Fixing bugs and regression tests.  SimplifyModels should be ready to merge back into master
In addition to some fairly routine bugfixes:

== New Libraries ==
 - Added slf4j support via logback.  See Compiler.scala for an example of logging.  Configure with src/main/resources/logback.xml
 - Upgraded to a newer version of specs2

== Type Inference ==
The type inference lens's interactions with the typechecker and related parts of the system still call to mind a brown bear trying to dance on a basketball while a cat pokes at the canary rolling around inside.  Basically: Type-inference makes liberal use of TAny, which is fairly useless for anything downstream.  The best guess cache, for example, needs to know what type to make the data field.  We need a more 'nondeterminism-aware typechecker'  For now, however... the hack is to sprinkle InlineVGTerms around into places where you *really* need type information.