CTExplainer now doesn't panic when it gets a query with a WHERE clause.
Some of the new provenance code was a bit jury-rigged...
`provenance.Provenance.filterForToken` rewrites an operator to return precisely one row, the one for the specified provenance token.  Turns out that `provenance.Tracer`, a class intended to extract attribute-level provenance was removing all of the Select operations meant to filter the inputs down to one row.

Solution: Tracer now does both steps: filtering down to one row and tracing attribute-level provenance.  It's still a hack that'll need to be thoroughly rewritten for Aggregates, but not until we figure out the rewrite semantics that we want there.