Implementing a Key Repair Lens and tangential bugfixes
Re-added a new lens modeled after the MayBMS RepairKey operator (this was part of our ICDE Submission branch) (closes #123).
This lens is similar to the classical DISTINCT operator, but allows you to force a DISTINCT on a subset of the table's columns (essentially creating a new key) rather than on full set of columns.

To see the point, let's look at a few examples:

* If the source data has no duplicates at all, the RepairKey lens is a no-op.
* If the source data has duplicates, but the entire row is duplicated, the RepairKey lens is identical to a SELECT DISTINCT *

Otherwise, the Key Repair lens "collapses" the remaining columns.  Take the following data as an example.
```
A,B,C
1,2,3
1,3,1
2, ,1
1,2,
1,4,2
2,2,1
4,2,4
```
Let's say we create a lens with `KEY_REPAIR(A)`.  Although `A` is a key in the input, it is a key in the output.
```
A,B,C
1,?,?
2,2,1
4,2,4
```
* There is only one row where A=4, so this row is preserved intact
* There are two rows where A=2, but...
    * In both rows, C=1 (i.e., There is a functional dependency between A and C), so we have one distinct value for C.
    * In one of the two rows, B is null, so again we only have one distinct non-null value for B.
* There's a mess of rows where A = 1, so B and C are chosen essentially at random from among the possibilities: B is one of 2, 3, 4; and C is one of 1, 2, 3

Bugfixes/Tweaks:
- (Change to Database.update) It's now possible to have bare attribute names in lens arguments.  Should propagate this change to other lenses that take column or keyword arguments.
- (Change to Typechecker) Typechecker now correctly computes the schema of an aggregate
- (Change to ModelManager) Added a new "RequiresDatabase" trait that automates RequiresDatabaseReconnect
- (Change to PushdownSelections) Select(Aggregate(...)) no longer crashes the compiler
- (Change to SQLiteCompat) Added a new "FIRST" aggregate mirroring https://wiki.postgresql.org/wiki/First/last_(aggregate)