Added a "Hint" parameter to VGTerms
There are a number of cases where Models are required to run nested queries to get additional information.
For example, the WekaModel needs other attribute values from the same row.  In principle, these values are all present, but they're not accessible to the model itself.  Similarly, the KeyRepair lens needs other possible values for the same key.

The hints field is distinct from the args field, as it is not used to uniquely identify the variable being created by the VGTerm.  However, in cases where it is feasible to pass additional information to the model to accelerate it, the hints field may be used to do so.

As of right now, the hints field is not yet being used, but it is present and available.