Playing around with a more aggressive expression compiler
It's not ready for primetime, but for your consideration...
EvalInlined is a complied analog to the interpreted Eval.  EvalInlined uses scala lambdas as
a sort of ghetto bytecode: Expressions are reduced to a function T => PrimitiveValue, where T
is an arbitrary, opaque type that defines a scope for the expression to evaluate in.

This has the potential to significantly boost performance on the CureTests