Starting to play around with using Mimir to drive a plotting engine
So I got fed up with having to fire up python | ruby | etc... every time I want to do some analytics and generate some graphs.  This is stuff that we should be dogfooding Mimir with.  It's not on the critical path, but I'm starting to play around with 'Vegas', a visualization tool for Scala.  It's not quite working yet, but again, since it's not on the critical path I'm going to leave it in there.

For now, I can only get Vegas to open up an empty window.  More poking in a bit.

Also I found a bug in LoadCSV: It freaks Mimir out if you try to load a CSV with an empty header field (i.e., if you try to create a table column without a name).