Continuing the reorg to mitigate issues with ResultIterator
ResultIterator is only valid for use within the db.query(){ ... } block.  The reason for
this is that we need to ensure that close() is called, or we start getting resource leaks.
Unfortunately, scala is a little too efficient with its lists.  For example: sequence
conversion (toSeq) is done lazilly.  This means that even if you convert the resultSet to
a sequence, you'll *still* get a result that doesn't get populated until it's used.
The only way to fix this is to use (toIndexedSeq) or do a blank foreach() loop to force
every element to be materialized.

It's not a complete fix, but there's now a Seq type (ResultSeq) that provides
ResultSet-like functionality, but with materialized data.  Some other utility methods
are also provided on top of it (e.g., pull out a specific column).  Calling toSeq on it
now does the right thing(tm), although unfortunately that won't be the case if you call
toSeq on its derivatives (e.g., what you get from map{} or filter{}).