Undoing a RowId extraction optimization intended for performance reasons.
At some point in TupleBundler, I modified JDBCUtilities to produce a special extractor for rowids for SQLite, which uses integers.  This neglected the possibility of merging rowids together --- such rowids are no longer just integers.  There may be some value in revisiting the issue somewhere down the line: It might be nice to have something like a `Pure` rowid that we can actually treat as an integer, but for now I just undid it.  This fixes several of the test cases