Compiler reorg and post-VLDB cleanup
Decomposing some of the hacks written up for VLDB into a slightly more
elegant, modular approach.  Specifically, the package mimir.exec.mode
now contains a bunch of derivatives of CompilerMode, which is a fully
encapsulated description of one particular approach to compiling queries.
The specific module can be passed in to db.compiler.compile or db.query,
and it should do the right thing.