CSV DataSource Error Warnings
Add a csv datasource that will add two columns, one boolean column that is true if there is an error for that row and another that has the original raw line from the CSV if there is an error.  Add an adaptive schema that projects off the additional columns and creates cooresponding datawarnings for each var in the error row. Add a test case that demonstrates and tests usage of the DataSourceErrors adaptive schema.