Error-aware parser no longer chokes if Spark optimizes supplemental columns out
Spark datasources get two schemas as input: a dataSchema and a requiredSchema.  The expectation
is that the datasource produce rows matching requiredSchema.  This allows Spark to optimize
execution, informing the data source that some columns don't need to be read from the source.

It seems like the existing CSV data source assumes that none of its source columns will ever get
optimized out.  Whether correct or not, it seems like the supplemental columns that get added by
the error-aware parser *can* get optimized out if they are unused.  The data source now
explicitly checks which rows need to be added before doing so.