All test cases pass
The final batch had a multitude of small issues.
- A few issues with the Sparsity parser were uncovered after switching to it for expression parsing.  These have been pushed into sparsity and should show up in 1.5.  Simple workarounds have been added Mimir-side
- The rather unusual case of [NULL < const] wasn't being handled by EvalInline.  This was resolved by adding an optimization rule to replace NULL < const with NULL.
- Provenance wasn't properly propagating ROWIDs for LeftOuterJoin.  This has been fixed.

CureScenario *does* work, but works painfully slow at the moment (it sucks up 8GB of memory and growing).  I don't want the default tests to require a local spark cluster, so for the time being I've disabled it in the normal test run.

I've also removed a number of test cases that have historically been a pain to update with small changes.  We had a whole bevy of tests that differ only in e.g., a function name, and thus don't really test anything interesting.  These were also very painful to update for small changes in the parser or the expression AST, so I've cut them.

Most of the other fixes were just adding schema information and/or propagating the new column names throughout queries.