updates to enable building for scala 2.12
all of the tests are passing but one of the libraries we are using for google sheets spark datasource did not have a scala 2.12 version available, so i forked it and made some updates to enable scala 2.12 and created a pull request.  we either need to wait until that gets merged and published or publish our own version.