Simplifying Compiler to push more work into Spark
At one point, Mimir ran on SQLite.  Calls to Scala UDFs were sloooow and it made sense to do more UDF work in Mimir.  This is no longer the case with Spark.

Compiler.compile was designed to strip off easily evaluable operations from the top of queries.  Specifically: Unions, Projections, and Aggregates would be evaluated 100% within Mimir, while the backend processed joins, etc...  While this made sense with SQLite, it is now a significant disadvantage with Spark, as Spark can process the data more reliably.  This commit removes this "feature" of Mimir.