Cleaning up pathing for spark integration.
MimirSpark no longer refers to ivy or m2 for jar paths.  Instead, it consults the
current classpath (which may or may not refer to ivy, m2, bootstrap, or anything else).
We expect all relevant jars to appear on the classpath, so this should work in general.
The one exception is mimir-core, which will not appear as a JAR when mimir is run in
testing mode.  Here, we dynamically create a JAR file for the current run and load it
instead.  This is a bit waste, but I don't have a cleaner solution for now... at least
not without going whole hog back to internal spark.